ENHANCED: Portability, Prolog driver, many details
Now uses a single (SWI-)Prolog program for the overall management,
running the Prolog system(s), joining the data and generating the
chart.

We now have a system in place to support pretty much every Prolog
system, with or without modules.  SWI, Yap and SICStus can do the
job with the original sources.  For e.g., Scryer we transform the
sources and test beforehand which files Scryer can handle.